Added doughnut png image
If you dont dig this I can totally find something else. As for the background colored square, i chose a random color, we can change it. I tried making a circular canvas for the plate but 1) i didn't know how 2) I didn't know how it would affect the dimensions/collision things. But we can try it Monday if you'd like. I'm going to work on making the sprinkles bounce off of the doughnut's outside edge.